use threaded mocked server for testing
